requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1
Django              3.2.9
djangorestframework 3.12.4
flake8              4.0.1
mccabe              0.6.1
pip                 21.3.1
pycodestyle         2.8.0
pyflakes            2.4.0
pytz                2021.3
setuptools          58.2.0
sqlparse            0.4.2
wheel               0.37.0
